Removal of S3 methods from SleeperProperties
