+++ java/configuration/src/main/java/sleeper/configuration/properties/SleeperProperties.java
@@ -15,7 +15,6 @@
  */
 package sleeper.configuration.properties;
 
-import com.amazonaws.services.s3.AmazonS3;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.slf4j.Logger;
@@ -46,7 +45,6 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static java.util.function.Predicate.not;
-import static sleeper.configuration.properties.PropertiesUtils.loadProperties;
 
 /**
  * Holds values for Sleeper configuration properties. Abstract class which backs both
@@ -196,16 +194,6 @@ public abstract class SleeperProperties<T extends SleeperProperty> implements Sl
         properties.clear();
         properties.putAll(newProperties);
         init();
-    }
-
-    protected void saveToS3(AmazonS3 s3Client, String bucket, String key) {
-        LOGGER.debug("Uploading config to bucket {}", bucket);
-        s3Client.putObject(bucket, key, saveAsString());
-    }
-
-    protected void loadFromS3(AmazonS3 s3Client, String bucket, String key) {
-        String propertiesString = s3Client.getObjectAsString(bucket, key);
-        resetAndValidate(loadProperties(propertiesString));
     }
 
     public Stream<Map.Entry<String, String>> getUnknownProperties() {
