+++ java/configuration/src/main/java/sleeper/configuration/properties/instance/InstanceProperties.java
@@ -110,7 +110,7 @@ public class InstanceProperties extends SleeperProperties<InstanceProperty> {
     }
 
     public void loadFromS3(AmazonS3 s3Client, String bucket) {
-        super.loadFromS3(s3Client, bucket, S3_INSTANCE_PROPERTIES_FILE);
+        this.loadFromS3(s3Client, bucket, S3_INSTANCE_PROPERTIES_FILE);
     }
 
     public static Properties loadPropertiesFromS3GivenInstanceId(AmazonS3 s3Client, String instanceId) {
@@ -123,7 +123,7 @@ public class InstanceProperties extends SleeperProperties<InstanceProperty> {
     }
 
     public void saveToS3(AmazonS3 s3Client) {
-        super.saveToS3(s3Client, get(CONFIG_BUCKET), S3_INSTANCE_PROPERTIES_FILE);
+        this.saveToS3(s3Client, get(CONFIG_BUCKET), S3_INSTANCE_PROPERTIES_FILE);
         LOGGER.info("Saved instance properties to bucket {}, key {}", get(CONFIG_BUCKET), S3_INSTANCE_PROPERTIES_FILE);
     }
 
@@ -161,5 +161,15 @@ public class InstanceProperties extends SleeperProperties<InstanceProperty> {
             count++;
         }
         return builder.toString();
+    }
+
+    protected void saveToS3(AmazonS3 s3Client, String bucket, String key) {
+        LOGGER.debug("Uploading config to bucket {}", bucket);
+        s3Client.putObject(bucket, key, saveAsString());
+    }
+
+    protected void loadFromS3(AmazonS3 s3Client, String bucket, String key) {
+        String propertiesString = s3Client.getObjectAsString(bucket, key);
+        resetAndValidate(loadProperties(propertiesString));
     }
 }
